Implement deleting entries with load
Magento treats NULL entries in configuration as existing values and does
not fall back to using configuration from a higher scope. In most cases,
this is not the intended behaviour. Therefore, if a value is null,
delete it from configuration instead of saving a null. This still allows
saving empty configuration values by passing in an empty string.